Add `Html.Lazy` to the top-level triangle.
This saves re-rendering the entire triangle for every single frame, only
requiring a full rerender when a node is "activated" or the text
changes. Note that this could _probably_ be used recursively too, to
optimize the tree further down the line.

However, just adding it as a wrapper was the simplest change with a
significant impact.